Fix the readers and writers count issue.
